models/BLIP/requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==8.0.3
datasets==2.12.0
decorator==4.4.2
deepspeed==0.5.1
distlib==0.3.0
fairscale==0.4.4
huggingface-hub==0.14.1
idna==2.9
joblib==1.1.0
jsonlines==3.0.0
numpy==1.22.2
opencv-python==4.5.1.48
packaging==21.3
pathtools==0.1.2
Pillow==9.0.1
prompt-toolkit==3.0.5
protobuf==3.19.4
psutil==5.8.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==6.0
regex==2022.1.18
requests==2.23.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.47
timm==0.4.12
tokenizers==0.10.3
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
transformers==4.15.0
typing-extensions==4.0.1
urllib3==1.25.9
wandb==0.12.10
